# src/rest/api/v2/languages/dto/language-with-id.ts
import { ApiProperty, PartialType } from '@nestjs/swagger';

import LanguageDto from './language.dto';

export class LanguageWithIdDto extends PartialType(LanguageDto) {
  @ApiProperty()
  dp_id: number;
}